+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added CHANGELOG.md.
 - Added base setup for Animation Nodes' extension.
 - Added build and stale workflows.
+- Added *Material Instancer* node.
 
 ### Fixed